requirements.txt
flask==2.2.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
pandas==2.1.1
numpy==1.24.3
plotly==5.18.0
scikit-learn==1.3.0
scipy==1.11.3
eventlet==0.33.3
python-jose==3.3.0
werkzeug==2.2.3 